Add Surface Mesh Deformer script
This version takes the topology of the mesh into account and calculates the distances based on the surface distance between vertices.

This will seem weird for meshes with discontinuous topology (unity sphere meshes!), but will work well for things like character meshes where elements like feet may be located closely to each other in euclidean space, but far in the geodesic distance on the surface of the mesh.